refactor: Extract ProgressBar component
